Move segmentation loading into what was previously the scale page
